requirements.txt
fastapi
uvicorn
telethon
python-dotenv
pydantic
firebase-admin